Simplified Near Cache code samples.

+++ clients/client-near-cache/src/main/java/NearCacheWithMemoryFormatBinary.java
@@ -9,17 +9,17 @@ public class NearCacheWithMemoryFormatBinary extends NearCacheSupport {
         HazelcastInstance member = Hazelcast.newHazelcastInstance();
         HazelcastInstance client = HazelcastClient.newHazelcastClient();
 
-        IMap<Long, Article> map = member.getMap("articlesBinary");
+        IMap<Integer, Article> map = member.getMap("articlesBinary");
         Article article = new Article("foo");
-        map.put(1L, article);
+        map.put(1, article);
 
         map = client.getMap("articlesBinary");
 
         // the first get() will populate the Near Cache
-        Article firstGet = map.get(1L);
+        Article firstGet = map.get(1);
         // the second and third get() will be served from the Near Cache
-        Article secondGet = map.get(1L);
-        Article thirdGet = map.get(1L);
+        Article secondGet = map.get(1);
+        Article thirdGet = map.get(1);
 
         printNearCacheStats(map);
