+++ clients/client-near-cache/src/main/java/NearCacheWithMemoryFormatObject.java
@@ -9,17 +9,17 @@ public class NearCacheWithMemoryFormatObject extends NearCacheSupport {
         HazelcastInstance member = Hazelcast.newHazelcastInstance();
         HazelcastInstance client = HazelcastClient.newHazelcastClient();
 
-        IMap<Long, Article> map = member.getMap("articlesObject");
+        IMap<Integer, Article> map = member.getMap("articlesObject");
         Article article = new Article("foo");
-        map.put(1L, article);
+        map.put(1, article);
 
         map = client.getMap("articlesObject");
 
         // the first get() will populate the Near Cache
-        Article firstGet = map.get(1L);
+        Article firstGet = map.get(1);
         // the second and third get() will be served from the Near Cache
-        Article secondGet = map.get(1L);
-        Article thirdGet = map.get(1L);
+        Article secondGet = map.get(1);
+        Article thirdGet = map.get(1);
 
         printNearCacheStats(map);
